Ignore the combined sample counts when performing per-lane QC.
Unfortunately with the ugrade from R 3.1.1 to 3.2.0, function heatmap.2
from the gplots package now extends much farther into the negative
range. Since the adjusted correlation coefficients are at their smallest
a little under zero, this makes the plots very ugly. I confirmed with
the old data that this is not a problem with the new data processing.